feat($add): add real number calculation logic
실수 덧셈 로직 추가:
- 커스텀 구분자로 .이 들어갈 시에는 정수 연산, 아닐 경우에는 소수점으로 취급하여 실수 연산 진행

+++ src/main/java/calculator/model/Numbers.java
@@ -13,10 +13,10 @@ public class Numbers {
         return new Numbers(numbers);
     }
 
-    public int add() {
-        int sum = 0;
+    public double add() {
+        double sum = 0;
         for (String number : numbers) {
-            sum += Integer.parseInt(number);
+            sum += Double.parseDouble(number);
         }
         return sum;
     }
